refactor: 半角→全角変換処理とモーラ数カウント処理を移動
