+++ crates/voicevox_core/src/engine/talk/text.rs
@@ -1,0 +1,2 @@
+pub(crate) mod hankaku_zenkaku;
+pub(super) mod katakana;
