+++ crates/voicevox_core/src/engine/talk/user_dict/word.rs
@@ -5,8 +5,11 @@ use serde::{de::Error as _, Deserialize, Serialize, Serializer};
 
 use crate::{error::ErrorRepr, result::Result};
 
-use super::part_of_speech_data::{
-    priority2cost, PartOfSpeechDetail, MAX_PRIORITY, MIN_PRIORITY, PART_OF_SPEECH_DETAIL,
+use super::{
+    super::text::{hankaku_zenkaku, katakana},
+    part_of_speech_data::{
+        priority2cost, PartOfSpeechDetail, MAX_PRIORITY, MIN_PRIORITY, PART_OF_SPEECH_DETAIL,
+    },
 };
 
 /// ユーザー辞書の単語。
@@ -179,18 +182,6 @@ pub const DEFAULT_PRIORITY: u32 = 5;
 
 static PRONUNCIATION_REGEX: LazyLock<Regex> =
     LazyLock::new(|| Regex::new(r"^[ァ-ヴー]+$").unwrap());
-static MORA_REGEX: LazyLock<Regex> = LazyLock::new(|| {
-    Regex::new(concat!(
-        "(?:",
-        "[イ][ェ]|[ヴ][ャュョ]|[トド][ゥ]|[テデ][ィャュョ]|[デ][ェ]|[クグ][ヮ]|", // rule_others
-        "[キシチニヒミリギジビピ][ェャュョ]|",                                    // rule_line_i
-        "[ツフヴ][ァ]|[ウスツフヴズ][ィ]|[ウツフヴ][ェォ]|",                      // rule_line_u
-        "[ァ-ヴー]",                                                              // rule_one_mora
-        ")",
-    ))
-    .unwrap()
-});
-static SPACE_REGEX: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"\p{Z}").unwrap());
 
 impl UserDictWord {
     #[cfg_attr(doc, doc(alias = "voicevox_user_dict_word_make"))]
@@ -211,7 +202,7 @@ impl UserDictWord {
         validate_pronunciation(&pronunciation)?;
         let mora_count = calculate_mora_count(&pronunciation, accent_type)?;
         Ok(Self {
-            surface: to_zenkaku(surface),
+            surface: hankaku_zenkaku::to_zenkaku(surface),
             pronunciation,
             accent_type,
             word_type,
@@ -290,7 +281,7 @@ pub(crate) fn validate_pronunciation(pronunciation: &str) -> InvalidWordResult<(
 /// カタカナの発音からモーラ数を計算する。
 fn calculate_mora_count(pronunciation: &str, accent_type: usize) -> InvalidWordResult<usize> {
     // 元実装：https://github.com/VOICEVOX/voicevox_engine/blob/39747666aa0895699e188f3fd03a0f448c9cf746/voicevox_engine/model.py#L212-L236
-    let mora_count = MORA_REGEX.find_iter(pronunciation).count();
+    let mora_count = katakana::count_moras(pronunciation);
 
     if accent_type > mora_count {
         return Err(InvalidWordError::InvalidAccentType(
@@ -300,26 +291,6 @@ fn calculate_mora_count(pronunciation: &str, accent_type: usize) -> InvalidWordR
     }
 
     Ok(mora_count)
-}
-
-/// 一部の種類の文字を、全角文字に置き換える。
-///
-/// 具体的には
-///
-/// - "!"から"~"までの範囲の文字(数字やアルファベット)は、対応する全角文字に
-/// - " "などの目に見えない文字は、まとめて全角スペース(0x3000)に
-///
-/// 変換する。
-pub(crate) fn to_zenkaku(surface: &str) -> String {
-    // 元実装：https://github.com/VOICEVOX/voicevox/blob/69898f5dd001d28d4de355a25766acb0e0833ec2/src/components/DictionaryManageDialog.vue#L379-L387
-    SPACE_REGEX
-        .replace_all(surface, "\u{3000}")
-        .chars()
-        .map(|c| match u32::from(c) {
-            i @ 0x21..=0x7e => char::from_u32(0xfee0 + i).unwrap_or(c),
-            _ => c,
-        })
-        .collect()
 }
 
 impl UserDictWordBuilder {
@@ -433,15 +404,6 @@ mod tests {
     use serde_json::json;
 
     use super::{InvalidWordError, UserDictWord, UserDictWordType};
-
-    #[rstest]
-    #[case("abcdefg", "ａｂｃｄｅｆｇ")]
-    #[case("あいうえお", "あいうえお")]
-    #[case("a_b_c_d_e_f_g", "ａ＿ｂ＿ｃ＿ｄ＿ｅ＿ｆ＿ｇ")]
-    #[case("a b c d e f g", "ａ　ｂ　ｃ　ｄ　ｅ　ｆ　ｇ")]
-    fn to_zenkaku_works(#[case] before: &str, #[case] after: &str) {
-        assert_eq!(super::to_zenkaku(before), after);
-    }
 
     #[rstest]
     fn to_mecab_format_works() {
